+++ app/Http/Controllers/WorkflowInstanceController.php
@@ -18,16 +18,27 @@ class WorkflowInstanceController extends Controller
     }
     public function index()
     {
-        $instances = WorkflowInstance::with(['definition', 'starter', 'updater', 'completer'])
-            ->orderBy('started_at', 'desc')
-            ->paginate(20);
+        $query = WorkflowInstance::with(['definition', 'starter', 'updater', 'completer'])
+            ->orderBy('started_at', 'desc');
+
+        if (!Auth::user()->isSuperAdmin()) {
+            $query->byOrganisation(Auth::user()->current_organisation_id);
+        }
+
+        $instances = $query->paginate(20);
 
         return view('workflows.instances.index', compact('instances'));
     }
 
     public function create()
     {
-        return view('workflows.instances.create', ['definitions' => WorkflowDefinition::active()->get()]);
+        $query = WorkflowDefinition::active();
+
+        if (!Auth::user()->isSuperAdmin()) {
+            $query->byOrganisation(Auth::user()->current_organisation_id);
+        }
+
+        return view('workflows.instances.create', ['definitions' => $query->get()]);
     }
 
     public function store(Request $request)
@@ -45,6 +56,7 @@ class WorkflowInstanceController extends Controller
             ...$validated,
             'status' => 'running',
             'current_state' => [],
+            'organisation_id' => Auth::user()->current_organisation_id,
             'started_by' => Auth::id(),
         ]);
 
@@ -54,12 +66,14 @@ class WorkflowInstanceController extends Controller
 
     public function show(WorkflowInstance $instance)
     {
+        $this->authorize('view', $instance);
         $instance->load(['definition', 'tasks', 'starter']);
         return view('workflows.instances.show', compact('instance'));
     }
 
     public function destroy(WorkflowInstance $instance)
     {
+        $this->authorize('delete', $instance);
         $instance->delete();
 
         return redirect()->route('workflows.instances.index')
@@ -71,6 +85,8 @@ class WorkflowInstanceController extends Controller
      */
     public function start(WorkflowInstance $instance)
     {
+        $this->authorize('update', $instance);
+
         try {
             $this->workflowEngine->startWorkflow($instance);
 
@@ -87,6 +103,8 @@ class WorkflowInstanceController extends Controller
      */
     public function pause(WorkflowInstance $instance)
     {
+        $this->authorize('update', $instance);
+
         try {
             $this->workflowEngine->pauseWorkflow($instance);
 
@@ -103,6 +121,8 @@ class WorkflowInstanceController extends Controller
      */
     public function resume(WorkflowInstance $instance)
     {
+        $this->authorize('update', $instance);
+
         try {
             $this->workflowEngine->resumeWorkflow($instance);
 
@@ -119,6 +139,8 @@ class WorkflowInstanceController extends Controller
      */
     public function cancel(WorkflowInstance $instance)
     {
+        $this->authorize('update', $instance);
+
         try {
             $this->workflowEngine->cancelWorkflow($instance);
